Added error checking for company Dirs, todos

+++ server/ftp_controller.py
@@ -16,14 +16,18 @@ def delete_files_from_dir(dir):
 def get_file_full(file_name, company_id):
     session = FTP('145.24.222.235') #Create session with FTP
     session.login("Controller", "cC2G'Q_&3qY@=D!@")
+
+    if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
+        return {"errorCode": 404, "Message": "Company directory does not exist on FTP server"}
+
     session.cwd(f'{company_id}') #Change directory on FTP to the company's
-    #TODO: Catch errors if doesnt exist
 
     random_file_path = generate_random_path(24, 'html') #Generate random file path for temp storage
     if path.exists(f'temporary_ftp_storage/{random_file_path}'): #Check for extreme edge case, if path is same as a different parallel request path
         random_file_path = generate_random_path(24, 'html')
 
     with open(f'temporary_ftp_storage/{random_file_path}', 'w') as handle: #Create the local (temporary) file
+        #TODO: Check if file exists on FTP server
         session.retrlines("RETR " + file_name, handle.write) #Retrieve file from FTP server and write it to created temp file
 
     return_data = io.BytesIO() #Create variable to store the local file contents
@@ -63,9 +67,15 @@ def upload_file(file_path, file_name, file_type, company_id):
     session.quit() #Close FTP session
     file_to_send.close()
 
-def delete_file(file_name, company_id=1):
+def delete_file(file_name, company_id):
     session = FTP('145.24.222.235')
     session.login("Controller", "cC2G'Q_&3qY@=D!@")
+
+    if f"{company_id}" not in session.nlst(): #Check if company dir exists on FTP Server, if not, return
+        return {"errorCode": 404, "Message": "Company directory does not exist on FTP server"}
+
+    session.cwd(f'{company_id}') #Change to the company dir
+    #TODO: Check if file exists in correct Dir
     session.delete(file_name)
     session.quit()
